# Implement `vec_type_common()` in C