--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -21,7 +21,7 @@ Imports:
     backports,
     digest,
     glue,
-    rlang (>= 0.3.3),
+    rlang (>= 0.3.4.9000),
     zeallot
 Suggests: 
     bit64,
@@ -34,6 +34,8 @@ Suggests:
     rmarkdown,
     testthat,
     tibble
+Remotes:
+    r-lib/rlang
 VignetteBuilder: 
     knitr
 Encoding: UTF-8