--- a/R/partial-frame.R
+++ b/R/partial-frame.R
@@ -67,6 +67,14 @@ vec_type2.vctrs_partial_frame <- function(x, y) {
   UseMethod("vec_type2.vctrs_partial_frame", y)
 }
 
+#' @method vec_type2.vctrs_partial_frame vctrs_partial_frame
+#' @export
+vec_type2.vctrs_partial_frame.vctrs_partial_frame <- function(x, y) {
+  partial <- vec_type2(x$partial, y$partial)
+  learned <- vec_type2(x$learned, y$learned)
+  new_partial_frame(partial, learned)
+}
+
 #' @method vec_type2.vctrs_partial_frame data.frame
 #' @export
 vec_type2.vctrs_partial_frame.data.frame <- function(x, y) {