--- a/R/type.R
+++ b/R/type.R
@@ -68,7 +68,7 @@ vec_type <- function(x) {
 #' @export
 #' @rdname vec_type
 vec_type_common <- function(..., .ptype = NULL) {
- .Call(vctrs_type_common, dots_values(.ptype, ...), .ptype)
+ .External2(vctrs_ext2_type_common, .ptype)
 }
 
 #' @export