--- a/src/init.c
+++ b/src/init.c
@@ -44,7 +44,6 @@ extern SEXP vec_proxy(SEXP);
 extern SEXP vctrs_unspecified(SEXP);
 extern SEXP vec_type(SEXP);
 extern SEXP vec_type_finalise(SEXP);
-extern SEXP vctrs_type_common(SEXP, SEXP);
 
 // Defined below
 SEXP vctrs_init(SEXP);
@@ -87,13 +86,18 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_unspecified",                (DL_FUNC) &vctrs_unspecified, 1},
   {"vctrs_type",                       (DL_FUNC) &vec_type, 1},
   {"vctrs_type_finalise",              (DL_FUNC) &vec_type_finalise, 1},
-  {"vctrs_type_common",                (DL_FUNC) &vctrs_type_common, 2},
   {NULL, NULL, 0}
+};
+
+extern SEXP vctrs_ext2_type_common(SEXP, SEXP, SEXP, SEXP);
+
+static const R_ExternalMethodDef ExtEntries[] = {
+  {"vctrs_ext2_type_common",           (DL_FUNC) &vctrs_ext2_type_common, 1},
 };
 
 void R_init_vctrs(DllInfo *dll)
 {
-    R_registerRoutines(dll, NULL, CallEntries, NULL, NULL);
+    R_registerRoutines(dll, NULL, CallEntries, NULL, ExtEntries);
     R_useDynamicSymbols(dll, FALSE);
 }