--- a/src/utils.c
+++ b/src/utils.c
@@ -4,6 +4,7 @@
 // Initialised at load time
 bool (*rlang_is_splice_box)(SEXP) = NULL;
 SEXP (*rlang_unbox)(SEXP) = NULL;
+SEXP (*rlang_env_dots_values)(SEXP) = NULL;
 
 
 bool is_bool(SEXP x) {
@@ -330,4 +331,5 @@ void vctrs_init_utils(SEXP ns) {
 
   rlang_is_splice_box = (bool (*)(SEXP)) R_GetCCallable("rlang", "rlang_is_splice_box");
   rlang_unbox = (SEXP (*)(SEXP)) R_GetCCallable("rlang", "rlang_unbox");
+  rlang_env_dots_values = (SEXP (*)(SEXP)) R_GetCCallable("rlang", "rlang_env_dots_values");
 }