--- a/src/utils.c
+++ b/src/utils.c
@@ -278,6 +278,19 @@ SEXP r_protect(SEXP x) {
   return Rf_lang2(fns_quote, x);
 }
 
+bool r_is_bool(SEXP x) {
+  return TYPEOF(x) == LGLSXP &&
+    Rf_length(x) == 1 &&
+    LOGICAL(x)[0] != NA_LOGICAL;
+}
+bool r_is_true(SEXP x) {
+  return r_is_bool(x) && LOGICAL(x)[0] == 1;
+}
+
+SEXP r_peek_option(const char* option) {
+  return Rf_GetOption1(Rf_install(option));
+}
+
 
 SEXP vctrs_ns_env = NULL;