--- a/src/utils.h
+++ b/src/utils.h
@@ -24,6 +24,7 @@ R_len_t compact_rownames_length(SEXP x);
 
 bool (*rlang_is_splice_box)(SEXP);
 SEXP (*rlang_unbox)(SEXP);
+SEXP (*rlang_env_dots_values)(SEXP);
 
 R_len_t r_lgl_sum(SEXP lgl, bool na_true);
 SEXP r_lgl_which(SEXP x, bool na_true);