--- a/src/utils.h
+++ b/src/utils.h
@@ -22,6 +22,9 @@ enum vctrs_type2 vec_typeof2_impl(enum vctrs_type type_x, enum vctrs_type type_y
 bool is_compact_rownames(SEXP x);
 R_len_t compact_rownames_length(SEXP x);
 
+bool (*rlang_is_splice_box)(SEXP);
+SEXP (*rlang_unbox)(SEXP);
+
 R_len_t r_lgl_sum(SEXP lgl, bool na_true);
 SEXP r_lgl_which(SEXP x, bool na_true);