Uses `rlang::dots_values()` for zero-cost splicing of lists. The common type finder recurses with the splice box holding the spliced values. These are standardised by rlang and always coerced to a list, possibly via `as.list()`.

Uses `.External2()` to invoke the C routine as it's faster and passes the current environment directly.

```r
### Few inputs

bench::mark(
  before = vec_type_common_before(1.5, 1:10),
  after = vec_type_common(1.5, 1:10)
)[1:6]
#>   expression      min     mean   median      max `itr/sec`
#>   <chr>      <bch:tm> <bch:tm> <bch:tm> <bch:tm>     <dbl>
#> 1 before      20.27µs  25.82µs  23.76µs  138.6µs    38722.
#> 2 after        1.55µs   1.98µs   1.82µs   31.5µs   506108.


### Few inputs - spliced

few <- rep(list(1.5, 1:10), 2)
bench::mark(
  before = vec_type_common_before(!!!few),
  after = vec_type_common(!!!few)
)[1:6]
#>   expression      min     mean   median      max `itr/sec`
#>   <chr>      <bch:tm> <bch:tm> <bch:tm> <bch:tm>     <dbl>
#> 1 before      24.19µs  31.32µs  28.75µs  146.9µs    31934.
#> 2 after        1.52µs   1.84µs   1.72µs   16.7µs   544413.


### Many inputs

many <- rep(list(1.5, 1:10), 1000)
bench::mark(
  before = vec_type_common_before(!!!x),
  after = vec_type_common(!!!x)
)[1:6]
#>   expression      min     mean   median      max `itr/sec`
#>   <chr>      <bch:tm> <bch:tm> <bch:tm> <bch:tm>     <dbl>
#> 1 before       3.95ms   4.48ms   4.48ms   5.38ms      223.
#> 2 after       81.39µs   93.1µs   87.5µs 250.73µs    10741.
```